fix: Set default Content-Type for POST request
If a POST request does not have a Content-Type set and the body has a size greater than zero, then an entry is added to the headers with type application/x-www-form-urlencoded. This was not being done if the body was an InputStream, so add a fix to ensure it is.